Fix arithmetic overflow issues in search engine
- Fix time_check_node_counter overflow: use wrapping_add to prevent panic
- Fix alpha/beta calculation overflow: use saturating_add/sub when calculating aspiration window bounds
- Fix window size calculation overflow: clamp f64 results to i32 range before casting in calculate_dynamic_window_size, calculate_adaptive_window_size, and calculate_comprehensive_window_size
- Fix depth squared multiplication overflow: cast depth to wider type (i32/u32) before squaring in search_engine.rs and move_ordering.rs to prevent u8 overflow when depth > 16
- Fix unused variable warning: remove unused engine_path return value in revalidate_engine_metadata
- Update engine path resolution to prefer debug binary in development mode

+++ src/search/move_ordering.rs
@@ -4406,7 +4406,9 @@ impl MoveOrdering {
     pub fn update_history_score(&mut self, move_: &Move, depth: u8) {
         if let Some(from) = move_.from {
             let key = (move_.piece_type, from, move_.to);
-            let bonus = (depth * depth) as u32; // Bonus proportional to depth
+            // Use safe multiplication to prevent overflow (depth is u8, max value is 255)
+            // depth * depth can overflow u8 if depth > 16, so cast to u32 first
+            let bonus = ((depth as u32) * (depth as u32)) as u32; // Bonus proportional to depth
             
             let current_score = self.history_table.get(&key).copied().unwrap_or(0);
             let new_score = current_score + bonus;
